fixed problem with modifiernames containing a colon
Pseudo classes are always starting with a colon as the first character. But the current regex searches for any colon in the modifier name, which causes problems with modifiers that contain a colon. One example would be the following CSS rule: `[data-u-customAttr~='align:left']` where the modifier `align:left` would gets falsely modified.