updated Handlebars Generator
There is an issue if you try to initiate the kssHandlebarsGenerator via the JS-API directly. If no custom helpers are defined the script is throwing an error because in line 109 the check this.config.helpers.length > 0 will throw an error if this.config.helpers = undefined .
This small change should fix it.