Allow to place "Styleguide" property anywhere
On my case, I need to place a custom property as last property of documentation. Without this fix, the sections with this custom property are not passed as sections on the styleguide object and are not rendered.

I propose this solution to fix the issue.

Just one question: Is the 'styleguide' property really required to be the last property of the documentation?