appropriately hide svgs from ATs
we don’t want keyboard/assistive technology users to focus within the SVGs, on some browsers, and we don’t need the SVGs announced, as there is accessible text associated with the toolbar child elements. So an aria-hidden=“true” and focusable=“false” has been added to the SVGs to mitigate these issues.